Moved Turn Group Utility
